Move charts out of widget tree
