+++ lib/data/podcast_genre.dart
@@ -1,0 +1,42 @@
+import 'package:musicpod/string_x.dart';
+
+enum PodcastGenre {
+  arts,
+  business,
+  comedy,
+  education,
+  fiction,
+  government,
+  healthAndFitness,
+  history,
+  kidsAndFamily,
+  leisure,
+  music,
+  news,
+  religionAndSpirituality,
+  science,
+  societyAndCulture,
+  sports,
+  tvAndFilm,
+  technology,
+  trueCrime;
+
+  String get id {
+    switch (this) {
+      case healthAndFitness:
+        return 'Health & Fitness';
+      case kidsAndFamily:
+        return 'Kids & Family';
+      case religionAndSpirituality:
+        return 'Religion & Spirituality';
+      case societyAndCulture:
+        return 'Society & Culture';
+      case tvAndFilm:
+        return 'TV & Film';
+      case trueCrime:
+        return 'True Crime';
+      default:
+        return name.capitalize();
+    }
+  }
+}
